refactor: isolate bootstrap priority imports to prevent circular dependencies
